refactor(app): mobile
