+++ client/app/app.less
@@ -2,7 +2,6 @@
 @import './style/variables.less';
 @import './style/buttons.less';
 @import './style/form.less';
-@import './style/chevron.less';
 
 [ng\:cloak], [ng-cloak], [data-ng-cloak], [x-ng-cloak], .ng-cloak, .x-ng-cloak {
   display: none !important;
