+++ client/app/common/components/header-bar/header-bar.component.less
