+++ client/app/common/components/header-bar/images/xively-logo.png
