+++ client/app/common/components/header-bar/index.js
@@ -1,0 +1,3 @@
+const headerBar = require('./header-bar.module')
+
+module.exports = headerBar.name
