--- client/app/common/components/header/header.module.js
@@ -1,7 +1,0 @@
-const angular = require('angular')
-const headerComponent = require('./header.component')
-
-const header = angular.module('simulator.common.components.header', [])
-  .component('header', headerComponent)
-
-module.exports = header
