--- client/app/common/components/header/index.js
@@ -1,3 +1,0 @@
-const header = require('./header.module')
-
-module.exports = header.name
