+++ client/app/simulator/mobile/images/arrow-down.png
