+++ client/app/simulator/mobile/index.js
@@ -1,0 +1,3 @@
+const mobileModule = require('./mobile.module')
+
+module.exports = mobileModule.name
