+++ client/app/simulator/mobile/timeseries-chart.component.js
