+++ client/app/simulator/mobile/timeseries-chart.component.less
@@ -1,4 +1,4 @@
-timeseries-chart, .timeseries-chart {
+timeseries-chart {
   display: block;
   width: 100%;
   height: 300px;
