+++ client/app/simulator/simulator.module.js
@@ -1,12 +1,16 @@
 const angular = require('angular')
+const angularAnimate = require('angular-animate')
+const angularRouter = require('angular-ui-router')
 
 const commonModule = require('../common')
+const mobileModule = require('./mobile')
 const simulatorRoute = require('./simulator.route')
 
 const simulatorModule = angular.module('simulator.devies', [
-  require('angular-animate'),
-  require('angular-ui-router'),
-  commonModule
+  angularAnimate,
+  angularRouter,
+  commonModule,
+  mobileModule
 ])
   .config(simulatorRoute)
