--- client/app/style/chevron.less
@@ -1,19 +1,0 @@
-.chevron::before {
-  border-style: solid;
-  border-width: 0.25em 0.25em 0 0;
-  content: '';
-  display: inline-block;
-  height: 0.45em;
-  left: 0.15em;
-  position: relative;
-  top: 0.15em;
-  transform: rotate(45deg);
-  vertical-align: top;
-  width: 0.45em;
-  transition: 0.1s all linear;
-}
-
-.chevron.bottom:before {
-  top: 0;
-  transform: rotate(135deg);
-}
